application-code/ecsdemo-cicd/.dagger/pyproject.toml
[project]
name = "ecsdemo-cicd"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.36.8",
    "dagger-io",
]

[tool.uv.sources]
dagger-io = { path = "sdk", editable = true }

[build-system]
requires = ["hatchling==1.25.0"]
build-backend = "hatchling.build"
